source/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py=2.1.0=pypi_0
alabaster=0.7.16=pypi_0
annotated-types=0.6.0=pypi_0
anthropic=0.25.1=pypi_0
anyio=4.3.0=pypi_0
arrow=1.3.0=pypi_0
attrs=23.2.0=pypi_0
babel=2.14.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
bibtexparser=1.4.1=pypi_0
blas=1.0=mkl
blinker=1.7.0=pypi_0
bottleneck=1.3.7=py311hd7041d2_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
bzip2=1.0.8=he774522_0
ca-certificates=2024.3.11=haa95532_0
cachetools=5.3.3=pypi_0
certifi=2024.2.2=pypi_0
cffi=1.16.0=pypi_0
charset-normalizer=3.3.2=pypi_0
click=8.1.7=py311haa95532_0
colorama=0.4.6=py311haa95532_0
contourpy=1.2.0=py311h59b6b97_0
cycler=0.11.0=pyhd3eb1b0_0
deprecated=1.2.14=pypi_0
distro=1.9.0=pypi_0
docstring-parser=0.16=pypi_0
docutils=0.20.1=pypi_0
fake-useragent=1.4.0=pypi_0
filelock=3.13.4=pypi_0
flask=3.0.2=pypi_0
fonttools=4.25.0=pyhd3eb1b0_0
free-proxy=1.1.1=pypi_0
freetype=2.12.1=ha860e81_0
fsspec=2024.3.1=pypi_0
google-ai-generativelanguage=0.6.1=pypi_0
google-api-core=2.18.0=pypi_0
google-api-python-client=2.125.0=pypi_0
google-auth=2.29.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-cloud-aiplatform=1.47.0=pypi_0
google-cloud-bigquery=3.20.1=pypi_0
google-cloud-core=2.4.1=pypi_0
google-cloud-resource-manager=1.12.3=pypi_0
google-cloud-storage=2.16.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-resumable-media=2.7.0=pypi_0
googleapis-common-protos=1.63.0=pypi_0
grpc-google-iam-v1=0.13.0=pypi_0
grpcio=1.62.1=pypi_0
grpcio-status=1.62.1=pypi_0
h11=0.14.0=pypi_0
httpcore=1.0.3=pypi_0
httplib2=0.22.0=pypi_0
httpx=0.26.0=pypi_0
huggingface-hub=0.23.3=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.6=pypi_0
imagesize=1.4.1=pypi_0
iniconfig=2.0.0=pypi_0
intel-openmp=2021.4.0=pypi_0
itsdangerous=2.1.2=pypi_0
jinja2=3.1.3=pypi_0
joblib=1.2.0=py311haa95532_0
jpeg=9e=h2bbff1b_1
kiwisolver=1.4.4=py311hd77b12b_0
lerc=3.0=hd77b12b_0
levenshtein=0.25.1=pypi_0
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libdeflate=1.17=h2bbff1b_1
libffi=3.4.4=hd77b12b_0
libpng=1.6.39=h8cc25b3_0
libtiff=4.5.1=hd77b12b_0
libwebp-base=1.3.2=h2bbff1b_0
lxml=5.1.0=pypi_0
lz4-c=1.9.4=h2bbff1b_0
markupsafe=2.1.5=pypi_0
matplotlib-base=3.8.0=py311hf62ec03_0
mkl=2021.4.0=pypi_0
mkl-service=2.4.0=py311h2bbff1b_1
mkl_fft=1.3.8=py311h2bbff1b_0
mkl_random=1.2.4=py311h59b6b97_0
mpmath=1.3.0=pypi_0
munkres=1.1.4=py_0
networkx=3.3=pypi_0
nltk=3.8.1=py311haa95532_0
numexpr=2.8.7=py311h1fcbade_0
numpy=1.26.3=pypi_0
numpy-base=1.24.3=py311hd01c5d8_1
openai=1.12.0=pypi_0
openjpeg=2.4.0=h4fc8c34_0
openssl=3.0.13=h2bbff1b_0
outcome=1.3.0.post0=pypi_0
packaging=23.2=py311haa95532_0
pandas=2.2.1=py311hea22821_0
patsy=0.5.3=py311haa95532_0
pillow=10.2.0=py311h2bbff1b_0
pinecone-client=3.2.1=pypi_0
pip=23.3.1=py311haa95532_0
pluggy=1.4.0=pypi_0
proto-plus=1.23.0=pypi_0
protobuf=4.25.3=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycparser=2.21=pypi_0
pydantic=2.6.1=pypi_0
pydantic-core=2.16.2=pypi_0
pygments=2.17.2=pypi_0
pyparsing=3.1.1=pypi_0
pysocks=1.7.1=pypi_0
pytest=8.0.0=pypi_0
python=3.11.7=he1021f5_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-dotenv=1.0.1=pypi_0
python-levenshtein=0.25.1=pypi_0
python-tzdata=2023.3=pyhd3eb1b0_0
pytz=2024.1=pypi_0
pyyaml=6.0.1=pypi_0
rapidfuzz=3.9.3=pypi_0
regex=2023.10.3=py311h2bbff1b_0
requests=2.31.0=pypi_0
rouge-score=0.1.2=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.3=pypi_0
scholarly=1.7.11=pypi_0
scikit-learn=1.3.0=py311hf62ec03_1
scikit-posthocs=0.9.0=pypi_0
scipy=1.11.4=py311hc1ccb85_0
seaborn=0.12.2=py311haa95532_0
selenium=4.18.1=pypi_0
sentence-transformers=3.0.0=pypi_0
setuptools=68.2.2=py311haa95532_0
shapely=2.0.3=pypi_0
six=1.16.0=pyhd3eb1b0_1
sniffio=1.3.0=pypi_0
snowballstemmer=2.2.0=pypi_0
sortedcontainers=2.4.0=pypi_0
soupsieve=2.5=pypi_0
sphinx=7.2.6=pypi_0
sphinx-rtd-theme=2.0.0=pypi_0
sphinxcontrib-applehelp=1.0.8=pypi_0
sphinxcontrib-devhelp=1.0.6=pypi_0
sphinxcontrib-htmlhelp=2.0.5=pypi_0
sphinxcontrib-jquery=4.1=pypi_0
sphinxcontrib-jsmath=1.0.1=pypi_0
sphinxcontrib-qthelp=1.0.7=pypi_0
sphinxcontrib-serializinghtml=1.1.10=pypi_0
sqlite=3.41.2=h2bbff1b_0
statsmodels=0.14.0=py311hd7041d2_0
sympy=1.12.1=pypi_0
tbb=2021.12.0=pypi_0
threadpoolctl=2.2.0=pyh0d69192_0
tiktoken=0.7.0=pypi_0
tk=8.6.12=h2bbff1b_0
tokenizers=0.19.1=pypi_0
torch=2.3.1=pypi_0
tqdm=4.66.2=pypi_0
transformers=4.41.2=pypi_0
trio=0.24.0=pypi_0
trio-websocket=0.11.1=pypi_0
types-python-dateutil=2.8.19.20240106=pypi_0
typing-extensions=4.9.0=pypi_0
tzdata=2024.1=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.0=pypi_0
vc=14.2=h21ff451_1
vs2015_runtime=14.27.29016=h5e58377_2
werkzeug=3.0.1=pypi_0
wheel=0.41.2=py311haa95532_0
wrapt=1.16.0=pypi_0
wsproto=1.2.0=pypi_0
xz=5.4.5=h8cc25b3_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
